Add Powerup Parent as Prefab
Powerups have 2 colliders; 1 is for physics simulation and smaller, the
other collider is slightly larger and is considered a trigger.

Landing check is based on the Ground Y Position public value, which is
0.3 like the landing detection for Dango. It's very hacky but the other
way to do it is using raycasting; performance might not be problem on PC
but could be issue on mobile.

Powerups also disappear 2 seconds after landing. The duration can be
changed by the timeLandedToDestroy public value.

To test, move a bunch of PowerupParent prefab into the scene.